Move issue_134 tests to test_handle Add test_real_ prefix to assign_double and assign_int test names
