+++ tests/test_cases/test_cocotb/test_handle.py
@@ -4,7 +4,10 @@
 """
 Tests for handles
 """
+import logging
+import random
 import cocotb
+from cocotb.triggers import Timer
 from cocotb.result import TestFailure
 
 from common import assert_raises
@@ -65,3 +68,58 @@ async def test_delayed_assignment_still_errors(dut):
         dut.stream_in_int <= "1010 not a real binary string"
     with assert_raises(TypeError):
         dut.stream_in_int <= []
+
+
+@cocotb.test(expect_error=cocotb.SIM_NAME in ["Icarus Verilog"])
+def test_integer(dut):
+    """
+    Test access to integers
+    """
+    log = logging.getLogger("cocotb.test")
+    yield Timer(10)
+    dut.stream_in_int = 4
+    yield Timer(10)
+    yield Timer(10)
+    got_in = int(dut.stream_out_int)
+    got_out = int(dut.stream_in_int)
+    log.info("dut.stream_out_int = %d" % got_out)
+    log.info("dut.stream_in_int = %d" % got_in)
+    if got_in != got_out:
+        raise TestFailure("stream_in_int and stream_out_int should not match")
+
+
+@cocotb.test(expect_error=cocotb.SIM_NAME in ["Icarus Verilog"])
+def test_real_assign_double(dut):
+    """
+    Assign a random floating point value, read it back from the DUT and check
+    it matches what we assigned
+    """
+    val = random.uniform(-1e307, 1e307)
+    log = logging.getLogger("cocotb.test")
+    yield Timer(1)
+    log.info("Setting the value %g" % val)
+    dut.stream_in_real = val
+    yield Timer(1)
+    yield Timer(1)  # FIXME: Workaround for VHPI scheduling - needs investigation
+    got = float(dut.stream_out_real)
+    log.info("Read back value %g" % got)
+    if got != val:
+        raise TestFailure("Values didn't match!")
+
+
+@cocotb.test(expect_error=cocotb.SIM_NAME in ["Icarus Verilog"])
+def test_real_assign_int(dut):
+    """Assign a random integer value to ensure we can write types convertible to
+    int, read it back from the DUT and check it matches what we assigned.
+    """
+    val = random.randint(-2**31, 2**31 - 1)
+    log = logging.getLogger("cocotb.test")
+    yield Timer(1)
+    log.info("Setting the value %i" % val)
+    dut.stream_in_real <= val
+    yield Timer(1)
+    yield Timer(1)  # FIXME: Workaround for VHPI scheduling - needs investigation
+    got = dut.stream_out_real
+    log.info("Read back value %d" % got)
+    if got != float(val):
+        raise TestFailure("Values didn't match!")
